// roi-tool/roi-frontend/src/core/rbac/permissions.enum.ts
export enum Permission {
  ViewUsers = "ViewUsers",
  CreateUser = "CreateUser",
  EditUser = "EditUser",
  DeleteUser = "DeleteUser",

  ViewCases = "ViewCases",
  CreateCases = "CreateCases",
  EditCases = "EditCases",
  EditCaseDetails = "EditCaseDetails",
  AddCaseMembers = "AddCaseMembers",
  CloseCases = "CloseCases",
  ArchiveCases = "ArchiveCases",

  ViewPartitions = "ViewPartitions",
  CreatePartitions = "CreatePartitions",
  EditPartitions = "EditPartitions",

  ViewFiles = "ViewFiles",
  UploadFiles = "UploadFiles",
  DownloadFiles = "DownloadFiles",
  DeleteFiles = "DeleteFiles",

  ViewWorkflows = "ViewWorkflows",
  EditWorkflows = "EditWorkflows",

  ViewSettings = "ViewSettings",

  ViewDatasets = "ViewDatasets",
  EditDatasets = "EditDatasets",
}